+++ eth_defi/uniswap_v2/pair.py
@@ -146,7 +146,7 @@ def fetch_pair_details(
         assert reverse_token_order is None, f"Give either (base_token_address, quote_token_address) or reverse_token_order"
         reverse_token_order = int(base_token_address, 16) > int(quote_token_address, 16)
 
-    pool = get_deployed_contract(web3, "UniswapV2Pair.json", pair_contact_address)
+    pool = get_deployed_contract(web3, "sushi/UniswapV2Pair.json", pair_contact_address)
     token0_address = pool.functions.token0().call()
     token1_address = pool.functions.token1().call()