+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Current
 
-- Reorganise ABI compilation process, source dependencies and `eth_defi.abi` folder layout
+- Reorganise ABI compilation process, source dependencies and `eth_defi.abi` folder layout\
+- In-house contracts are now compiled using [Foundry](https://book.getfoundry.sh/)
 - Add `eth_defi.enzyme.vault_controlled_vallet`
+
 
 # 0.16.1